GROUND_GPS msg received in OCAML, problems with newly created RL_TELEPORT Message

+++ sw/airborne/modules/rl_soaring/rl_soaring.h
@@ -28,7 +28,7 @@
 
 #include "std.h"
 
-extern int rl_teleported;
+extern int16_t rl_teleported;
 
 typedef struct{
     // states
